Move coordinator for powerwall integration to separate module
Create a `PowerwallUpdateCoordinator` subclass of `DataUpdateCoordinator`
in a dedicated `coordinator.py` module, alongside `PowerwallDataManager`,
`PowerwallConfigEntry`, and `PowerwallRuntimeData`.
