+++ homeassistant/components/powerwall/entity.py
@@ -1,10 +1,7 @@
 """The Tesla Powerwall integration base entity."""
 
 from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-)
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import (
     DOMAIN,
@@ -14,10 +11,11 @@ from .const import (
     POWERWALL_BASE_INFO,
     POWERWALL_COORDINATOR,
 )
-from .models import BatteryResponse, PowerwallData, PowerwallRuntimeData
+from .coordinator import PowerwallRuntimeData, PowerwallUpdateCoordinator
+from .models import BatteryResponse, PowerwallData
 
 
-class PowerWallEntity(CoordinatorEntity[DataUpdateCoordinator[PowerwallData]]):
+class PowerWallEntity(CoordinatorEntity[PowerwallUpdateCoordinator]):
     """Base class for powerwall entities."""
 
     _attr_has_entity_name = True
@@ -45,7 +43,7 @@ class PowerWallEntity(CoordinatorEntity[DataUpdateCoordinator[PowerwallData]]):
         return self.coordinator.data
 
 
-class BatteryEntity(CoordinatorEntity[DataUpdateCoordinator[PowerwallData]]):
+class BatteryEntity(CoordinatorEntity[PowerwallUpdateCoordinator]):
     """Base class for battery entities."""
 
     _attr_has_entity_name = True
